- Textures, framebuffers  - Introduced more typedefs, makes it easier to change types when needed (size_t and ubyte for now)

+++ coffee/cregex.h
@@ -26,7 +26,7 @@ static std::vector<CRegexMatch> coffee_regex_match_multi(const CString& pattern,
         CRegexMatch m;
         m.b_match = false;
         if(std::regex_match(string,mch,rgx)){
-            for(size_t i=0;i<mch.size();i++){
+            for(uint32_t i=0;i<mch.size();i++){
                 std::ssub_match smch = mch[i];
                 m.b_match = true;
                 if(capture)
@@ -44,7 +44,7 @@ static CRegexMatch coffee_regex_match(CString pattern, const CString& data, bool
     CRegexMatch m;
     m.b_match = false;
     if(std::regex_match(data,mch,rgx)){
-        for(size_t i=0;i<mch.size();i++){
+        for(uint32_t i=0;i<mch.size();i++){
             std::ssub_match smch = mch[i];
             m.b_match = true;
             if(capture)
